+++ tests/init.py
@@ -11,7 +11,7 @@ from __future__ import annotations
 from typing import Any, Callable, Tuple, cast, Optional
 
 from urllib.parse import ParseResult, urlparse, parse_qs
-import sys, io, atexit
+import sys, io, atexit, base64
 import unittest
 
 from rich.console import Console
@@ -368,8 +368,24 @@ def sendRequest(operation:Operation, url:str, originator:str, ty:ResourceTypes=N
 		return None, 5103
 
 
+def addHttpAuthorizationHeader(headers:Parameters) -> Optional[Tuple[str, int]]:
+	global oauthToken
+
+	if doOAuth:
+		if (token := OAuth.getOAuthToken(oauthServerUrl, oauthClientID, oauthClientSecret, oauthToken)) is None:
+			return 'error retrieving oauth token', 5103
+		oauthToken = token
+		headers['Authorization'] = f'Bearer {oauthToken.token}'
+	elif doHttpBasicAuth:
+		_t = f'{httpUserName}:{httpPassword}'
+		headers['Authorization'] = f'Basic {base64.b64encode(_t.encode("utf-8")).decode("utf-8")}'
+	elif doHttpTokenAuth:
+		headers['Authorization'] = f'Bearer {httpAuthToken}'
+	return None
+
+
 def sendHttpRequest(method:Callable, url:str, originator:str, ty:ResourceTypes=None, data:JSON|str=None, ct:str=None, timeout:float=None, headers:Parameters=None) -> Tuple[STRING|JSON, int]:	# type: ignore # TODO Constants
-	global oauthToken, httpSession
+	global httpSession
 
 	# correct url
 	url = RequestUtils.toHttpUrl(url)
@@ -405,11 +421,19 @@ def sendHttpRequest(method:Callable, url:str, originator:str, ty:ResourceTypes=N
 		hds.update(headers)
 	
 	# authentication
-	if doOAuth:
-		if (token := OAuth.getOAuthToken(oauthServerUrl, oauthClientID, oauthClientSecret, oauthToken)) is None:
-			return 'error retrieving oauth token', 5103
-		oauthToken = token
-		hds['Authorization'] = f'Bearer {oauthToken.token}'
+	if (_r := addHttpAuthorizationHeader(hds)) is not None:
+		return _r
+
+	# if doOAuth:
+	# 	if (token := OAuth.getOAuthToken(oauthServerUrl, oauthClientID, oauthClientSecret, oauthToken)) is None:
+	# 		return 'error retrieving oauth token', 5103
+	# 	oauthToken = token
+	# 	hds['Authorization'] = f'Bearer {oauthToken.token}'
+	# elif doHttpBasicAuth:
+	# 	_t = f'{httpUserName}:{httpPassword}'
+	# 	hds['Authorization'] = f'Basic {base64.b64encode(_t.encode("utf-8")).decode("utf-8")}'
+	# elif doHttpTokenAuth:
+	# 	hds['Authorization'] = f'Bearer aRandomToken'
 
 	# Verbose output
 	if verboseRequests:
@@ -664,7 +688,9 @@ def enableShortResourceExpirations() -> None:
 	global _orgExpCheck, _maxExpiration, _tooLargeResourceExpirationDelta
 
 	# Send UT request
-	resp = requests.post(UTURL, headers = { UTCMD: f'enableShortResourceExpiration {expirationCheckDelay}'})
+	headers = { UTCMD: f'enableShortResourceExpiration {expirationCheckDelay}'}
+	addHttpAuthorizationHeader(headers)
+	resp = requests.post(UTURL, headers = headers)
 	_maxExpiration = -1
 	_orgExpCheck = -1
 	if resp.status_code == 200:
@@ -684,7 +710,9 @@ def disableShortResourceExpirations() -> None:
 	global _orgExpCheck, _orgREQExpCheck
 	if _orgExpCheck != -1:
 		# Send UT request
-		resp = requests.post(UTURL, headers = { UTCMD: f'disableShortResourceExpiration'})
+		headers = { UTCMD: f'disableShortResourceExpiration'}
+		addHttpAuthorizationHeader(headers)
+		resp = requests.post(UTURL, headers = headers)
 		if resp.status_code == 200:
 			_orgExpCheck = -1
 			_orgREQExpCheck = -1
@@ -716,7 +744,9 @@ def enableShortRequestExpirations() -> None:
 	global _orgRequestExpirationDelta
 
 	# Send UT request
-	resp = requests.post(UTURL, headers = { UTCMD: f'enableShortRequestExpiration {requestExpirationDelay}'})
+	headers = { UTCMD: f'enableShortRequestExpiration {requestExpirationDelay}'}
+	addHttpAuthorizationHeader(headers)
+	resp = requests.post(UTURL, headers = headers)
 	if resp.status_code == 200:
 		if UTRSP in resp.headers:
 			_orgRequestExpirationDelta = float(resp.headers[UTRSP])
@@ -728,7 +758,9 @@ def disableShortRequestExpirations() -> None:
 	global _orgRequestExpirationDelta
 	
 	# Send UT request
-	resp = requests.post(UTURL, headers = { UTCMD: f'disableShortRequestExpiration'})
+	headers = { UTCMD: f'disableShortRequestExpiration'}
+	addHttpAuthorizationHeader(headers)
+	resp = requests.post(UTURL, headers = headers)
 	if resp.status_code == 200:
 		_orgRequestExpirationDelta = -1.0
 	
@@ -749,7 +781,9 @@ def testCaseStart(name:str) -> None:
 			name: Name of the test case.
 	"""
 	if UPPERTESTERENABLED:
-		requests.post(UTURL, headers = { UTCMD: f'testCaseStart {name}'})
+		headers = { UTCMD: f'testCaseStart {name}'}
+		addHttpAuthorizationHeader(headers)
+		requests.post(UTURL, headers = headers)
 	if verboseRequests:
 		console.print('')
 		ln  = '=' * int((console.width - 11 - len(name)) / 2)
@@ -764,7 +798,9 @@ def testCaseEnd(name:str) -> None:
 			name: Name of the test case.
 	"""
 	if UPPERTESTERENABLED:
-		requests.post(UTURL, headers = { UTCMD: f'testCaseEnd {name}'})
+		headers = { UTCMD: f'testCaseEnd {name}'}
+		addHttpAuthorizationHeader(headers)
+		requests.post(UTURL, headers = headers)
 	if verboseRequests:
 		console.print('')
 		ln  = '=' * int((console.width - 9 - len(name)) / 2)
@@ -1115,10 +1151,20 @@ noRemote = not connectionPossible(REMOTEcseURL)
 
 if UPPERTESTERENABLED:
 	try:
-		if requests.post(UTURL, headers = { UTCMD: f'Status'}).status_code != 200:
-			console.print('[red]Upper Tester Interface not enabeled in CSE')
-			console.print('Enable with configuration setting: "\[http]:enableUpperTesterEndpoint=True"')
-			quit(-1)
+		headers = { UTCMD: f'Status'}
+		addHttpAuthorizationHeader(headers)
+		response = requests.post(UTURL, headers = headers)
+		match response.status_code:
+			case 200:
+				pass
+			case 401:
+				console.print('[red]CSE requires authorization')
+				console.print('Add authorization settings to the test suite configuration file')
+				quit(-1)
+			case _:
+				console.print('[red]Upper Tester Interface not enabeled in CSE')
+				console.print('Enable with configuration setting: "\[http]:enableUpperTesterEndpoint=True"')
+				quit(-1)
 	except (ConnectionRefusedError, requests.exceptions.ConnectionError):
 		console.print('[red]Connection to CSE not possible[/red]\nIs it running?')
 		quit(-1)